Adding Assignment 4 files
DESCRIPTION
Take a provided information architecture (taxonomy) in the form of an
HTML template with a list of links, to serve as a site navigation
widget.

INSTRUCTIONS
1. Using the supplied HTML template, containing an unordered list of
hyperlinks, build a simple CSS-powered drop-down menu.
2. Make sure that the second-level link lists are concealed, until the
first-level list item is hovered over.
3. Apply an external style sheet with a link element or an @import rule
and place all the CSS rules inside the file.
4. Apply some sort of interactivity with pseudo classes for the links
(:link, :visited, :hover,
:focus, :active).